Drop PMO specific parameters from config create command
While setting up the config, there are a bunch of DNS, hostname related commands
requested from the user. These fields are mandatory and will be prompted for
when the config create command is run. These options don't make much sense to be
asked in the PMK context. This change drops the prompts/mandatory property for
these optional parameters.

This is a quick workaround and not necessarily the right long term fix. But with
the immediate priority on PMK, putting this in place till we figure out the
right long term solution