Simplify the manage DNS resolver handling
With click callbacks, if it is dependent on a CLI option it should be processed
before the callback is invoked. The eager setting is used to handle this
generally. However, we had to drop the eager setting in the earlier commit and
this leads to an ordering problem that the ex eager options have to be processed
before the callback is invoked.

In this change, we don't depend on callbacks but make an explicit call to manage
DNS resolver from inside the command processing.