Initial cut of the CLI setup script
A setup script to prep the express CLI. Many of its current implementation is
heavily inspired by the pf9-express setup process. In short the script does
following
- Check the platform (currenly supporting only Ubuntu)
- Check if py3 is available (first preference to use this), else use py2
- Ensure the right pip is installed.
- Create a venv in /opt/pf9/cli based on py2/py3 method
- Install some pre-reqs like ansible, jq, shade
- Install express-cli in the venv
- Run express init
- Run express config and ask for the DU creds. We want a config to be created
right away for the end user.

These steps are optimized for PMK at the moment because that is the first target
for the CLI.

Note that there is a test setup flag (-t) which will pull express from
test.pypi.org. That is where we stage intermediate versions of the express-cli
before release.

It is not polished enough yet but this should be a good place for users to play
around with this.
- Has command echo for debugging purposes.
- Logging of the setup needs to be standardized
- Some command failure handling is missing.
- Sym link to the exec itself to make it part of the user's PATH is not done