Drop PMO specific parameters from config create command (#26)
* Drop PMO specific parameters from config create command

While setting up the config, there are a bunch of DNS, hostname related commands
requested from the user. These fields are mandatory and will be prompted for
when the config create command is run. These options don't make much sense to be
asked in the PMK context. This change drops the prompts/mandatory property for
these optional parameters.

This is a quick workaround and not necessarily the right long term fix. But with
the immediate priority on PMK, putting this in place till we figure out the
right long term solution

* Simplify the manage DNS resolver handling

With click callbacks, if it is dependent on a CLI option it should be processed
before the callback is invoked. The eager setting is used to handle this
generally. However, we had to drop the eager setting in the earlier commit and
this leads to an ordering problem that the ex eager options have to be processed
before the callback is invoked.

In this change, we don't depend on callbacks but make an explicit call to manage
DNS resolver from inside the command processing.

* Updated manage_resolvers callback to correctly handle not being called. Added a test to validate the config create command against the minimal options