Pass custom python env to get access to venv modules + Expose ansible logs in (#48)
output

When installed in the venv, certain modules will not be readily available in the
global space. Ansible needs to be told to use a venv as the interpreter for
certain commands. This change passes an extra var to Ansible to allow that.
Ansible consumes this in a manner deemed appropriate for it.

Also includes changes to show the Ansible logs output to the end user in case of
failures.