Add warning message for non en_US.UTF-8 locale
Issue: https://github.com/platform9/express-cli/issues/76

We do not want to fail the install altogether because there are
different ways to specify the locale string. This is a best effort
fix to warn the user about possible failure.
As long as the system locale isn't purely ASCII based, the CLI will work.