Use the region's URL to accomodate secondary regions
While the prep-node was already taking care of adding nodes to the right region
even if it was a secondary region, the create cluster and attach cluster were
assuming primary region only.

This change determines the region URL to use for any qbert operations like
create and attach cluster.

Testing done: Validated that node attach works with secondary regions